build.sh: add data partition to ISO image
Currently, the only way to integrate data in the installer ISO (SSH
keys, artifacts, etc.) is to integrate them before building the ISO.
Once generated, it is not possible to add or delete these data.

To fix this, this commit implements the addition of a 512M data partition to
the ISO image. Once the ISO is burnt, the data partition can be mounted
to add data to the installer, without needing to rebuild the ISO.

Signed-off-by: Paul Le Guen de Kerneizon <paul.leguendekerneizon@savoirfairelinux.com>